app/requirements.txt
Flask
flask_sqlalchemy
flask_restful
uuid
PyJWT
